// src/app/ViewModels/models/part-dispatch-kit-report.model.ts
export interface PartDispatchKitReportViewModel {

  srNo: number;

  poNumber?: string;

  poDate?: Date;

  submitToERPDate?: Date;

  poType?: string;

  companyName?: string;

  mobileNo?: string;

  dealerCode?: string;

  dealerCity?: string;

  dealerState?: string;

  locationCode?: string;

  locationName?: string;

  locationCity?: string;
}